backend_firered/requirements.txt
# FireRed-Image-Edit-1.0 backend. Python 3.11.
# diffusers from main for QwenImageEditPlusPipeline (FireRed)
diffusers @ git+https://github.com/huggingface/diffusers
transformers>=4.40.0
accelerate>=0.33.0
safetensors>=0.4.0
torch>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
pillow>=10.0.0
